1. goApi order 添加订单

+++ goApi/util/helper/response.go
@@ -2,8 +2,12 @@ package helper
 
 type Response struct {
 	Code    int16
-	ErrCode int32
+	ErrCode int64
 	Status  string
 	Msg     string
 	Data    interface{}
 }
+
+
+
+
